requirements.txt
fastapi[standard]
pytest
perplexityai
requests
pydantic-settings
